requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
pydantic==2.9.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
Flask